ARM: tegra: Update isl29028 compatible string to use isil vendor prefix
"isil" and "isl" prefixes are used at various locations inside the kernel
to reference Intersil corporation. This patch is part of a series fixing
those locations were "isl" is used in compatible strings to use the now
expected "isil" prefix instead (NASDAQ symbol for Intersil and most used
version).

Note: isl29028 is an I2C device so the patch does not in fact currently
depend on the introduction of "isil"-based compatible string in isl29028
driver because I2C core does not check the prefix yet.

Signed-off-by: Arnaud Ebalard <arno@natisbad.org>
Signed-off-by: Thierry Reding <treding@nvidia.com>

[ upstream commit: 62981239a61bfd5306737effc015deba26c0d198 ]