dt: bindings: update mvebu-mbus DT binding with new compatible properties
Contrary to what was originally thought, the Armada 375 and Armada 38x
MBus windows hardware block is *not* compatible with the Armada 370,
due to a difference in how window 13 is handled. It was rather
compatible with the Armada XP MBus hardware block.

However, the DTs for Armada 375 and Armada 38x encode the following
compatible string for MBus:

  compatible = "marvell,armada375-mbus", "marvell,armada370-mbus", "simple-bus";
  compatible = "marvell,armada380-mbus", "marvell,armada370-mbus", "simple-bus";

So, by extending the mvebu-mbus DT binding to also cover the
marvell,armada375-mbus and marvell,armada380-mbus compatible strings,
we can define a new behavior for those SoCs without changing the DT.

Therefore, this commit adds those two new compatible strings to the DT
binding documentation of mvebu-mbus. Note that it re-uses two existing
duplicated lines for the armada370-mbus and armadaxp-mbus compatible
strings.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Andrew Lunn <andrew@lunn.ch>

[ upstream commit: ddefbdb4c0df6bcb751971e2395e65daf26abb10 ]