Merge branch 'xen-netfront-next'
David Vrabel says:

====================
xen-netfront: refactor making Tx requests

As netfront as evolved to handle different sorts of skbs the code to
fill a Tx requests has been copy and pasted several times.  The series
refactors this and a few other areas.

The first patch is to a Xen header but this can be merged via
net-next.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 6a38cc2be61c2293112adfa019158b9dc6e7eb03 ]