Merge branch 'link_netns_advertise'
Nicolas Dichtel says:

====================
netns: advertise netns via netlink

The first patch of the series fix a bug of the previous series (present in
net-next only).
The rest of the series adds an attribute to advertise the peer netns for
rtnetlink messages where this information is needed by userland to be able to
interpret fully the received message.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: f4ac8292b09350868418983fc1b85a6c6e48a177 ]