scripts merged from upstream; disabled onboard rtc; removed LCD gpio node temporarily
Merged in the scripts folder from upstream. Makefile now functions properly.

The onboard RTC has been disabled to prevent the kernel from accidentally
using it. This forces the kernel to use the battery backed RTC found on
the cape.

The LCD GPIO node has been temporarily removed as there currently is no
kernel module that could use it. A task for the future will be to move the
userspace LCD driver to kernel space.